Refactor Token's page controller
We are refactoring this controller to follow the same pattern that we
have been using in the Address's page and Transaction's page since it
uses the same approach.

This way, we can use a generic helper to deal with tabs.
